# backend/src/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "CommentsApi",
    "description": "CRUD сервис для работы с комментариями"
  },
  "host": "http://localhost:5000",
  "basePath": "",
  "tags": [
    {
      "name": "Comments",
      "description": "Эндпоинт для получения комментариев"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/comments/": {
      "get": {
        "tags": [
          "Comments"
        ],
        "description": "Возвращает список всех параметров",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Comments"
        ],
        "description": "Создает новый комментарий",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Комментарий",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string",
                  "example": "Автор комментария"
                },
                "objectId": {
                  "type": "string",
                  "example": "Объект комментария"
                },
                "body": {
                  "type": "string",
                  "example": "Тело комментария"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/comments/filter": {
      "get": {
        "tags": [
          "Comments"
        ],
        "description": "Возвращает Список комментариев по фильтрам. Список должнен удовлетворять ограничениям по каждому полю (AND). Можно добавлять несколько значений каждого поля.",
        "parameters": [
          {
            "name": "objectId",
            "in": "query",
            "description": "Объкты комментария",
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "Автор комментария",
            "type": "string"
          },
          {
            "name": "from",
            "in": "query",
            "description": "Начало интервала (дата последней модификации)",
            "type": "string"
          },
          {
            "name": "to",
            "in": "query",
            "description": "Конец интервала (дата последней модификации)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/comments/pages": {
      "get": {
        "tags": [
          "Comments"
        ],
        "description": "Возвращает страницы список комментариев.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Количество элементов",
            "required": true,
            "type": "integer"
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "id комментария после которого надо взять limit элементов",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "В body хранятся все комментарии, в cursor-id последнего комментария из body"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/comments/{id}": {
      "get": {
        "tags": [
          "Comments"
        ],
        "description": "Возвращает комментарий по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID comment"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Comments"
        ],
        "description": "Удаляет комментарий по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID comment"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Comments"
        ],
        "description": "Обновляет комментарий",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Комментарий",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string",
                  "example": "Автор комментария"
                },
                "objectId": {
                  "type": "string",
                  "example": "Объект комментария"
                },
                "body": {
                  "type": "string",
                  "example": "Тело комментария"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}